+++ app/components/nav_bar/base_component.rb
@@ -21,7 +21,7 @@ module NavBar
       links << Link.new(t(".home"), root_path) if helpers.turbo_native_app?
       links << Link.new(t(".developers"), developers_path)
       links << Link.new(t(".pricing"), pricing_path)
-      links << Link.new(t(".about"), about_path)
+      links << Link.new(t(".affiliates"), affiliates_path)
       links
     end
   end