+++ app/models/concerns/businesses/has_online_profiles.rb
@@ -1,0 +1,10 @@
+module Businesses
+  module HasOnlineProfiles
+    extend ActiveSupport::Concern
+    include UrlAttribute
+
+    included do
+      url_attribute :website
+    end
+  end
+end