+++ app/models/developers.rb
@@ -1,0 +1,5 @@
+module Developers
+  def self.table_name_prefix
+    "developers_"
+  end
+end