+++ config/routes.rb
@@ -7,6 +7,7 @@ Rails.application.routes.draw do
     }
 
     resource :about, only: :show, controller: :about
+    resources :affiliates, only: %w[index create new], controller: "affiliates/registrations"
     resource :conduct, only: :show
     resource :home, only: :show
     resource :pricing, only: :show, controller: :pricing
@@ -14,6 +15,10 @@ Rails.application.routes.draw do
 
     resources :businesses, except: :destroy
     resources :specialties, only: :index
+
+    namespace :businesses do
+      resources :hiring_invoice_requests, only: [:new, :create]
+    end
 
     # /notifications/read must come before /notifications/:id.
     resources :read_notifications, only: [:index, :create], path: "/notifications/read"
@@ -33,11 +38,13 @@ Rails.application.routes.draw do
       resources :public_profiles, only: :new
     end
 
+    namespace :developers do
+      resources :celebration_package_requests, only: [:new, :create]
+    end
+
     get "developers/:id/:key", to: "developers#show", as: :developer_public
 
-    resource :hired, only: :show, controller: :hired do
-      resources :forms, only: [:new, :create], module: :hired
-    end
+    resource :hired, only: :show, controller: :hired
 
     namespace :hiring_agreement, module: :hiring_agreements do
       resource :terms, only: :show
@@ -69,7 +76,9 @@ Rails.application.routes.draw do
     resources :conversations, only: :index
     resources :specialties
     resources :transactions, except: :show
-    resources :users, only: [:index]
+    resources :users, only: [:index] do
+      resources :referrals, only: :index, module: :users
+    end
 
     namespace :conversations do
       resources :blocks, only: :index
@@ -80,6 +89,10 @@ Rails.application.routes.draw do
       resources :invisiblizes, only: :create, module: :businesses
     end
 
+    namespace :businesses do
+      resources :hiring_invoice_requests, only: [:index, :show]
+    end
+
     resources :developers, only: [] do
       resource :source_contributors, only: %i[create destroy], module: :developers
       resources :conversations, only: :index, controller: :developer_conversations
@@ -87,8 +100,8 @@ Rails.application.routes.draw do
       resources :invisiblizes, only: :create, module: :developers
     end
 
-    namespace :hired do
-      resources :forms, only: [:index, :show]
+    namespace :developers do
+      resources :celebration_package_requests, only: [:index, :show]
     end
 
     namespace :hiring_agreements do
@@ -96,6 +109,8 @@ Rails.application.routes.draw do
         resource :activation, only: %i[create destroy], module: :terms
       end
     end
+
+    resources :referrals, only: :index
   end
 
   namespace :api, defaults: {format: :json} do
@@ -120,9 +135,6 @@ Rails.application.routes.draw do
     resource :postmark, only: :create, controller: :postmark
     resource :revenuecat, only: :create, controller: :revenue_cat
   end
-
-  # TODO: Added temporary route to be updated after business hiring form changes are merged
-  get "business_hiring_form", to: redirect("")
 
   get "/sitemap.xml.gz", to: redirect("#{Rails.configuration.sitemaps_host}sitemaps/sitemap.xml.gz"), as: :sitemap
   get "robots.:format" => "robots#index"