+++ db/migrate/20220610193544_add_authentication_token_to_users.rb
@@ -1,4 +1,8 @@
 class AddAuthenticationTokenToUsers < ActiveRecord::Migration[7.0]
+  class User < ActiveRecord::Base
+    has_secure_token :authentication_token
+  end
+
   def change
     add_column :users, :authentication_token, :string
     User.find_each.map(&:regenerate_authentication_token)