+++ db/migrate/20230323132220_add_survey_request_notfications_to_businesses.rb
@@ -1,0 +1,5 @@
+class AddSurveyRequestNotficationsToBusinesses < ActiveRecord::Migration[7.0]
+  def change
+    add_column :businesses, :survey_request_notifications, :boolean, default: true
+  end
+end