+++ db/migrate/20230324090527_add_scheduling_link_to_developers.rb
@@ -1,0 +1,5 @@
+class AddSchedulingLinkToDevelopers < ActiveRecord::Migration[7.0]
+  def change
+    add_column :developers, :scheduling_link, :string
+  end
+end