+++ db/migrate/20230331210502_add_counter_cache_to_user_referrals.rb
@@ -1,0 +1,7 @@
+class AddCounterCacheToUserReferrals < ActiveRecord::Migration[7.0]
+  def change
+    add_column :users, :referrals_count, :integer, null: false, default: 0
+
+    User.find_each { |u| User.reset_counters(u.id, :referrals) }
+  end
+end